Move FilterHelper to ObjectModel
