+++ src/Microsoft.TestPlatform.Common/Resources/xlf/Resources.cs.xlf
@@ -264,11 +264,6 @@
         <target state="translated">Při vytváření rychlého filtru došlo k chybě.</target>
         <note />
       </trans-unit>
-      <trans-unit id="TestCaseFilterEscapeException">
-        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
-        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
-        <note></note>
-      </trans-unit>
     </body>
   </file>
 </xliff>
