+++ src/Microsoft.TestPlatform.Common/Resources/xlf/Resources.es.xlf
@@ -264,11 +264,6 @@
         <target state="translated">Error al crear el filtro rápido.</target>
         <note />
       </trans-unit>
-      <trans-unit id="TestCaseFilterEscapeException">
-        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
-        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
-        <note></note>
-      </trans-unit>
     </body>
   </file>
 </xliff>
