+++ src/Microsoft.TestPlatform.Common/Resources/xlf/Resources.ru.xlf
@@ -264,11 +264,6 @@
         <target state="translated">При создании быстрого (Fast) фильтра произошла ошибка.</target>
         <note />
       </trans-unit>
-      <trans-unit id="TestCaseFilterEscapeException">
-        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
-        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
-        <note></note>
-      </trans-unit>
     </body>
   </file>
 </xliff>
