+++ src/Microsoft.TestPlatform.Common/Resources/xlf/Resources.xlf
@@ -141,11 +141,6 @@
         <target state="new">An error occured while creating Fast filter.</target>
         <note></note>
       </trans-unit>
-      <trans-unit id="TestCaseFilterEscapeException">
-        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
-        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
-        <note></note>
-      </trans-unit>
     </body>
   </file>
 </xliff>
