+++ src/Microsoft.TestPlatform.ObjectModel/Resources/xlf/Resources.cs.xlf
@@ -612,6 +612,11 @@
         <target state="translated">Neplatné nastavení {0}. Očekával se alespoň jeden XmlAttribute: popisný název (friendlyName), identifikátor URI (uri) nebo kvalifikovaný název sestavení (assemblyQualifiedName).</target>
         <note />
       </trans-unit>
+      <trans-unit id="TestCaseFilterEscapeException">
+        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
+        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
+        <note></note>
+      </trans-unit>
     </body>
   </file>
 </xliff>
