+++ src/Microsoft.TestPlatform.ObjectModel/Resources/xlf/Resources.de.xlf
@@ -612,6 +612,11 @@
         <target state="translated">Ungültige Einstellungen "{0}". Erwartet wird mindestens ein XmlAttribute: "friendlyName", "uri" oder "assemblyQualifiedName".</target>
         <note />
       </trans-unit>
+      <trans-unit id="TestCaseFilterEscapeException">
+        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
+        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
+        <note></note>
+      </trans-unit>
     </body>
   </file>
 </xliff>
