+++ src/Microsoft.TestPlatform.ObjectModel/Resources/xlf/Resources.es.xlf
@@ -612,6 +612,11 @@
         <target state="translated">Configuración "{0}" no válida. Se esperaba al menos uno de los siguientes valores de XmlAttribute: friendlyName, uri y assemblyQualifiedName.</target>
         <note />
       </trans-unit>
+      <trans-unit id="TestCaseFilterEscapeException">
+        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
+        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
+        <note></note>
+      </trans-unit>
     </body>
   </file>
 </xliff>
