+++ src/Microsoft.TestPlatform.ObjectModel/Resources/xlf/Resources.fr.xlf
@@ -612,6 +612,11 @@
         <target state="translated">Paramètres '{0}' non valides. Au moins un des attributs XmlAttribute (friendlyName, uri ou assemblyQualifiedName) était attendu.</target>
         <note />
       </trans-unit>
+      <trans-unit id="TestCaseFilterEscapeException">
+        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
+        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
+        <note></note>
+      </trans-unit>
     </body>
   </file>
 </xliff>
