+++ src/Microsoft.TestPlatform.ObjectModel/Resources/xlf/Resources.ja.xlf
@@ -612,6 +612,11 @@
         <target state="translated">設定 '{0}' は無効です。friendlyName、uri、assemblyQualifiedName のうち少なくとも一つの XmlAttribute が必要です。</target>
         <note />
       </trans-unit>
+      <trans-unit id="TestCaseFilterEscapeException">
+        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
+        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
+        <note></note>
+      </trans-unit>
     </body>
   </file>
 </xliff>
