+++ src/Microsoft.TestPlatform.ObjectModel/Resources/xlf/Resources.ru.xlf
@@ -612,6 +612,11 @@
         <target state="translated">Недопустимые параметры "{0}". Ожидается как минимум один атрибут XmlAttribute: friendlyName, uri или assemblyQualifiedName.</target>
         <note />
       </trans-unit>
+      <trans-unit id="TestCaseFilterEscapeException">
+        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
+        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
+        <note></note>
+      </trans-unit>
     </body>
   </file>
 </xliff>
