+++ src/Microsoft.TestPlatform.ObjectModel/Resources/xlf/Resources.tr.xlf
@@ -612,6 +612,11 @@
         <target state="translated">Geçersiz ayarlar '{0}'. FriendlyName, URI veya AssemblyQualifiedName değişkenlerinden en az biri bekleniyordu.</target>
         <note />
       </trans-unit>
+      <trans-unit id="TestCaseFilterEscapeException">
+        <source>Filter string '{0}' includes unrecognized escape sequence.</source>
+        <target state="new">Filter string '{0}' includes unrecognized escape sequence.</target>
+        <note></note>
+      </trans-unit>
     </body>
   </file>
 </xliff>
